text-posts stored in the database, visible in scrollingFeed
