+++ .env
@@ -1,2 +1,2 @@
 ENERGY=20 GeV
-JDBC_DATABASE_URL=jdbc:postgresql://localhost/postgres?user=postgres&password=ignatius
+JDBC_DATABASE_URL=jdbc:postgresql://localhost/postgres?user=postgres&password=root
