--- src/main/java/com/Post.java
@@ -1,38 +1,0 @@
-package com.example;
-
-public class Post
-{
-    private String title;
-    private String description;
-    private String category;
-
-    public String getTitle()
-    {
-        return this.title;
-    }
-
-    public String getDescription()
-    {
-        return this.description;
-    }
-
-    public String getCategory()
-    {
-        return this.category;
-    }
-
-    public void setTitle(String s)
-    {
-        title = s;
-    }
-
-    public void setDescription(String s)
-    {
-        description = s;
-    }
-
-    public void setCategory(String s)
-    {
-        category = s;
-    }
-}
